// src/services/youtube/@types/types.ts
export interface RetryOptions {
  maxRetries: number;
  backoff: {
    inc: number;
    max: number;
  };
}

export type PipelineEndpoint = (...args: any[]) => Promise<any>;